Move buffered functions into their own ext trait
